# CRLF parsing micro-optimization